Update playwright tests primary category
Update playwright tests primary category

Update playwright tests primary category

Update playwright tests primary category
